requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiofiles==0.4.0           # via responder
aniso8601==6.0.0          # via graphene
apispec==2.0.2            # via responder
apistar==0.7.2            # via responder
attrs==19.1.0
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests, responder
click==7.0                # via apistar, uvicorn
docopt==0.6.2             # via responder
graphene==2.1.7           # via responder
graphql-core==2.2.1       # via graphene, graphql-relay, graphql-server-core
graphql-relay==2.0.0      # via graphene
graphql-server-core==1.1.1  # via responder
h11==0.8.1                # via uvicorn
httptools==0.0.13         # via uvicorn
idna==2.8                 # via requests
itsdangerous==1.1.0       # via responder
jinja2==2.10.1            # via apistar, responder
markupsafe==1.1.1         # via jinja2
marshmallow==2.19.5       # via responder
oauthlib==3.0.2           # via requests-oauthlib
parse==1.12.0             # via responder
promise==2.2.1            # via graphql-core, graphql-relay, graphql-server-core
python-multipart==0.0.5   # via responder
pyyaml==5.1.1             # via apistar, responder
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1  # via responder
requests==2.22.0          # via apistar, requests-oauthlib, requests-toolbelt, responder
responder==1.3.1
rfc3986==1.3.2            # via responder
rx==1.6.1                 # via graphql-core
six==1.12.0               # via graphene, graphql-core, graphql-relay, promise, python-multipart
starlette==0.10.7         # via responder
typesystem==0.2.4         # via apistar
urllib3==1.25.3           # via requests
uvicorn==0.8.4            # via responder
uvloop==0.12.2            # via responder, uvicorn
websockets==7.0           # via uvicorn
whitenoise==4.1.3         # via responder
